+++ apps/stoop-mobile/src/screens/ItemDetailScreen.js
@@ -20,7 +20,7 @@ import { useNavigation, useRoute } from '@react-navigation/native';
 
 import { ROUTES }                            from '../navigation.js';
 import { COLORS, SPACING, FONT_SIZES, RADII } from '../lib/theme.js';
-import { t }                                 from '../lib/i18n.js';
+import { t }                                 from '../lib/localisation.js';
 import { AvatarCircle }                      from '../components/AvatarCircle.js';
 import { AttachmentModal }                   from '../components/AttachmentModal.js';
 import { attachmentUri, timeAgo }            from '../lib/post.js';